backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyMySQL==1.1.1
cryptography==42.0.8
django-cors-headers==4.4.0
Pillow>=11.1.0
django-environ==0.11.2
celery==5.4.0
django-filter==24.3
drf-spectacular==0.27.2
